refactor!: combine two menu-bar mixins to a single one
