--- packages/menu-bar/src/vaadin-menu-bar-buttons-mixin.d.ts
@@ -1,22 +1,0 @@
-/**
- * @license
- * Copyright (c) 2019 - 2023 Vaadin Ltd.
- * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
- */
-import type { Constructor } from '@open-wc/dedupe-mixin';
-import type { ControllerMixinClass } from '@vaadin/component-base/src/controller-mixin.js';
-import type { ResizeMixinClass } from '@vaadin/component-base/src/resize-mixin.js';
-
-export declare function ButtonsMixin<T extends Constructor<HTMLElement>>(
-  base: T,
-): Constructor<ButtonsMixinClass> & Constructor<ControllerMixinClass> & Constructor<ResizeMixinClass> & T;
-
-export declare class ButtonsMixinClass {
-  protected readonly _buttons: HTMLElement[];
-
-  protected readonly _container: HTMLElement;
-
-  protected readonly _overflow: HTMLElement;
-
-  protected _hasOverflow: boolean;
-}
