+++ packages/menu-bar/test/typings/menu-bar.types.ts
@@ -2,6 +2,7 @@ import '../../vaadin-menu-bar.js';
 import type { ControllerMixinClass } from '@vaadin/component-base/src/controller-mixin.js';
 import type { FocusMixinClass } from '@vaadin/component-base/src/focus-mixin.js';
 import type { ResizeMixinClass } from '@vaadin/component-base/src/resize-mixin.js';
+import type { MenuBarMixinClass } from '../../src/vaadin-menu-bar-mixin.js';
 import type { MenuBarItem, MenuBarItemSelectedEvent } from '../../vaadin-menu-bar.js';
 
 const menu = document.createElement('vaadin-menu-bar');
@@ -11,6 +12,7 @@ const assertType = <TExpected>(actual: TExpected) => actual;
 assertType<ResizeMixinClass>(menu);
 assertType<ControllerMixinClass>(menu);
 assertType<FocusMixinClass>(menu);
+assertType<MenuBarMixinClass>(menu);
 
 menu.addEventListener('item-selected', (event) => {
   assertType<MenuBarItemSelectedEvent>(event);
